dev-requirements.txt
dbt-duckdb v0.8.1

